M-Files_15-7/requirement.txt
streamlit==1.29.0
pillow==9.4.0
python-docx==0.8.11
reportlab==4.0.4
